requirements.txt
python-telegram-bot==20.7
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
aiosqlite==0.19.0 